[admin] fix typo - logical AND

+++ functions/check-started/index.js
@@ -44,7 +44,7 @@ module.exports = async function (context, checkStartMessage) {
   // Delete the row in the preparedCheck table for live checks only - prevent pupils logging in again.]
   try {
     context.log(`isLiveCheck:${checkData.isLiveCheck} delete:${config.DeletePreparedChecksFromTableStorageOnCheckStarted}`)
-    if (checkData.isLiveCheck & config.DeletePreparedChecksFromTableStorageOnCheckStarted) {
+    if (checkData.isLiveCheck && config.DeletePreparedChecksFromTableStorageOnCheckStarted) {
       await azureStorageHelper.deleteFromPreparedCheckTableStorage(azureTableService, checkStartMessage.checkCode, context.log)
       context.log('check-started: SUCCESS: pupil check row deleted from preparedCheck table for', checkStartMessage.checkCode)
       await requestPupilStatusChange(checkData.pupil_id, checkData.checkCode)
